requirements.txt
requests
bs4
mysql.connector
